# Conflicts:
#	win32/build/config.w32

+++ ext/bz2/bz2_filter.c
@@ -307,7 +307,7 @@ static php_stream_filter_ops php_bz2_compress_ops = {
 
 /* {{{ bzip2.* common factory */
 
-static php_stream_filter *php_bz2_filter_create(const char *filtername, zval *filterparams, int persistent)
+static php_stream_filter *php_bz2_filter_create(const char *filtername, zval *filterparams, uint8_t persistent)
 {
 	php_stream_filter_ops *fops = NULL;
 	php_bz2_filter_data *data;
